+++ src/panels/lovelace/components/hui-entity-editor.ts
@@ -19,7 +19,7 @@ import Sortable, {
 import { fireEvent } from "../../../common/dom/fire_event";
 import "../../../components/entity/ha-entity-picker";
 import "../../../components/ha-icon-button";
-import { sortableStyles } from "../../../components/ha-sidebar-sort-styles";
+import { sortableStyles } from "../../../resources/ha-sortable-style";
 import { HomeAssistant } from "../../../types";
 import { EditorTarget } from "../editor/types";
 import { EntityConfig } from "../entity-rows/types";
